[#13406] Change internal instructions counter to 64 bits.

+++ arch/translate-all.c
@@ -157,7 +157,7 @@ static inline void gen_block_footer(TranslationBlock *tb)
     *gen_opc_ptr = INDEX_op_end;
 }
 
-static int get_max_instruction_count(CPUState *env, TranslationBlock *tb)
+static uint64_t get_max_instruction_count(CPUState *env, TranslationBlock *tb)
 {
     int instructions_count = size_of_next_block_to_translate > 0
         ? size_of_next_block_to_translate
